refactor(dm-tool): decompose InitiativeTracker into focused components
Splits the 821-line monolithic InitiativeTracker.tsx into an
initiative-tracker/ subdirectory with one file per concern, making
each piece independently navigable and testable.
